Refactored commands to seperate files
