--- kattis.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-python kattis.py $*
